improved error handling in numerous routes, extracted code to helper functions for numerous routes
