+++ controllers/apps.js
@@ -7,241 +7,109 @@ const ini = require('ini');
 // const path = require("path");
 const AXIOS = require('axios');
 const { compileFunction } = require("vm");
-const { getClusterMetrics } = require('./cluster_metrics');
+const { getClusterMetrics } = require('./clusterMetricsHelpers');
+const HttpError = require("../models/httpError");
+const { 
+  getManagerIP,
+  createDockerAPIConnection,
+  getIDForAppName,
+  appendAppNames,
+  combineServicesWithTheirTasks,
+  onlyMostRecentServiceTasks, 
+  combineTasksWithNodeMetrics,
+  removeExtraneousProperties,
+  filterForDesiredService,
+  checkForCanary } = require("./appsHelpers");
 
-function pad(value) {
-  value = String(value);
-  if (value.length < 2) {
-    return "0" + value;
-  } else {
-    return value;
+async function getServiceLogs(req, res, next) {
+  const managerIP = getManagerIP();
+
+  try {
+    let serviceNameToLog = req.params.appName;
+    let idForLogs = await getIDForAppName(serviceNameToLog, managerIP);
+
+    let result = await AXIOS.get(`http://${managerIP}:2375/services/${idForLogs}/logs?stdout=true&stderr=true`);
+    let logs = result.data;
+    res.send(logs);
+  } catch(err) {
+    console.log(err);
+    next(new HttpError("Unable to get logs", 404));
   }
 }
 
-function timeStamp() {
-  var date = new Date(Date.now());
-  return `${+date.getDate()}/${+(date.getMonth()+1)}/${+date.getFullYear()}-${pad(+date.getHours())}:${pad(+date.getMinutes())}:${pad(+date.getSeconds())}`;
-}
-
-function getManagerIP() {
-  const hosts = ini.parse(fs.readFileSync('./ansible/inventory/hosts', 'utf-8'));
-  return Object.keys(hosts.managers)[0].split(' ')[0];
-}
-
-function createDockerAPIConnection() {
+function listServices(req, res, next) {
   const managerIP = getManagerIP();
 
-  return new Docker({
-    host: managerIP,
-    port: process.env.DOCKER_PORT || 2375,
-    // ca: fs.readFileSync('ca.pem'),
-    // cert: fs.readFileSync('cert.pem'),
-    // key: fs.readFileSync('~/ssh-keys/ec2-docker.pem'),
-    version: 'v1.25' // required when Docker >= v1.13, https://docs.docker.com/engine/api/version-history/
+  let result = AXIOS.get(`http://${managerIP}:2375/services`);
+  result.then(success => {
+    result = success.data;
+    result = result.map(record => {
+      return {
+        serviceName: record.Spec.Name,
+        serviceID: record.ID,
+      };
+    });
+
+    result = appendAppNames(result);
+
+    res.send(result);
+  }).catch(err => {
+    console.log(err);
+    next(new HttpError("Error getting inventory", 404));
   });
 }
 
-module.exports = {
-  // https://docs.docker.com/engine/api/v1.37/#operation/ServiceLogs
+async function inspectService(req, res, next) {
+  let serviceName = req.params.appName;
+  const managerIP = getManagerIP();
+  let nodeMetrics;
+  let tasks;
+  
+  try {
+    nodeMetrics = await getClusterMetrics(managerIP);
+  } catch (err) {
+    next(new HttpError("Unable to get details for this service", 404));
+  }
 
-  async getServiceLogs(req, res, next) {
-    if (!fs.existsSync('./ansible/inventory/hosts')) { // if hosts file does not exist respond with 404
-      res.status(404).send("Manager node does not exist.");
-    }
-    const managerIP = getManagerIP();
+  try {
+    tasks = await AXIOS.get(`http://${managerIP}:2375/tasks`);
+    tasks = tasks.data;
+  } catch (err) {
+    console.log(err);
+    next(new HttpError("Unable to get details for this service", 404));
+  }
 
-    let serviceNameToLog = req.params.appName;
+  try {
+    let services = await AXIOS.get(`http://${managerIP}:2375/services`)
+    services = services.data;
 
-    try {
-      let result = await AXIOS.get(`http://${managerIP}:2375/services`);
-      let services = result.data;
-      let justNamesAndIDs = services.map(record => {
-        return {
-          serviceName: record.Spec.Name,
-          serviceID: record.ID,
-        };
-      });
-
-      let idForLogs = justNamesAndIDs.filter(record => {
-        console.log(record.serviceName, serviceNameToLog);
-        return record.serviceName === serviceNameToLog;
-      })[0].serviceID;
-
-      let result2 = await AXIOS.get(`http://${managerIP}:2375/services/${idForLogs}/logs?stdout=true&stderr=true`);
-      let logs = result2.data;
-      res.send(logs);
-    } catch(err) {
-      console.log(err);
-      res.send(err);
-    }
-  },
-
-  listServices(req, res, next) {
-    const managerIP = getManagerIP();
-
-    let result = AXIOS.get(`http://${managerIP}:2375/services`);
-    result.then(success => {
-      result = success.data;
-      result = result.map(record => {
-        return {
-          serviceName: record.Spec.Name,
-          serviceID: record.ID,
-        };
-      });
-
-      let regexForLastPart = /(_production|_canary)/
-      let regexForFirstPart = /.*(?=_)/
-
-      result = result.map(record => {
-        if (regexForLastPart.test(record.serviceName)) {
-          record.appName = record.serviceName.match(regexForFirstPart)[0]
-        }
-        return record;
-      })
-
-      res.send(result);
-    }).catch(err => {
-      console.error(err);
+    let singleService = filterForDesiredService(services, serviceName);
+    combineServicesWithTheirTasks(singleService, tasks);
+    singleService.forEach(service => {
+      combineTasksWithNodeMetrics(service.serviceTasks, nodeMetrics)
     });
-  },
+    onlyMostRecentServiceTasks(singleService);
+    removeExtraneousProperties(singleService);
 
-  async inspectService(req, res, next) {
-    const managerIP = getManagerIP();
-    let clusterMetrics;
-    
-    try {
-      let nodeMetrics = await getClusterMetrics(managerIP);
+    let hasCanary = checkForCanary(singleService);
+    let message = "For more information on app performance, visit the prometheus and grafana dashboards you have configured with SENTINEL METRICS, or inspect app logs with SENTINEL INSPECT LOGS. You can also view system level metrics for your compute instances with SENTINEL CLUSTER INSPECT.";
 
-      clusterMetrics = nodeMetrics;
-    } catch (err) {
-      console.log(err);
-    }
+    let response = {
+      hasCanary,
+      message,
+      data: singleService
+    };
+    res.send(response);
+  } catch (err) {
+    console.log(err);
+    next(new HttpError("Unable to get details for this service", 404));
+  }
+}
 
-    let serviceName = req.params.appName;
-    let regex = /.*(?=[_-])/;
-    // let serviceNameFirstPart = serviceName.match(regex)[0];
-
-    try {
-      let services = await AXIOS.get(`http://${managerIP}:2375/services`)
-      services = services.data;
-      services = services.filter(record => {
-        let firstPart = record.Spec.Name.match(regex)[0];
-        // return firstPart === serviceNameFirstPart;
-        return firstPart === serviceName;
-      });
-      services = services.map(record => {
-        return {
-          serviceName: record.Spec.Name,
-          serviceID: record.ID,
-          serviceReplicas: record.Spec.Mode.Replicated ? record.Spec.Mode.Replicated.Replicas : 0,
-          servicesTasks: []
-        };
-      });
-
-      let tasks = await AXIOS.get(`http://${managerIP}:2375/tasks`);
-      tasks = tasks.data;
-
-      services.forEach(service => {
-        tasks.forEach(task => {
-          if (task.ServiceID === service.serviceID) {
-            service.servicesTasks.push({
-              taskStatus: task.Status.State,
-              taskStatusTimestamp: task.Status.Timestamp,
-              taskSlot: task.Slot,
-              taskContainer: task.Status.ContainerStatus.ContainerID,
-              taskNodeID: task.NodeID
-            });
-          };
-        });
-      });
-
-      services.forEach(service => {
-        service.servicesTasks.sort((a, b) => {
-          if (a.taskStatusTimestamp < b.taskStatusTimestamp) {
-            return 1
-          } else if (a.taskStatusTimestamp > b.taskStatusTimestamp) {
-            return -1
-          } else {
-            return 0
-          }
-        })
-      });
-
-      //console.log("THE METRICS: ", clusterMetrics)
-      services.forEach(service => {
-        service.servicesTasks = service.servicesTasks.slice(0, service.serviceReplicas);
-        service.servicesTasks.forEach(task => {
-          clusterMetrics.forEach(metric => {
-            if (task.taskNodeID === metric.NodeID) {
-              task.hostNodeMetrics = {
-                diskSpace: metric.DiskSpace,
-                memorySpace: metric.MemorySpace,
-                cpuUsageAvgLast10Minutes: metric.cpuUsageAvgLast10Minutes
-              }
-            }
-          })
-          task.taskStatusTimestamp = timeStamp(task.taskStatusTimestamp);
-          delete task.taskNodeID;
-          delete task.taskContainer;
-        })
-          delete service.serviceID;
-      })
-
-      
-
-      let result = services;
-
-      // let anyNotRunning = false;
-      // result.forEach(service => {
-      //   service.servicesTasks.forEach(task => {
-      //     if (task.taskStatus !== "running") {
-      //       anyNotRunning = true;
-      //     }
-      //   })
-      // });
-
-      let message = "For more information on app performance, visit the prometheus and grafana dashboards you have configured with SENTINEL METRICS, or inspect app logs with SENTINEL INSPECT LOGS. You can also view system level metrics for your compute instances with SENTINEL CLUSTER INSPECT.";
-      let justNames = result.map(record => {
-        return record.serviceName;
-      })
-      let regexCanary = /_canary/;
-
-      console.log(justNames);
-      let hasCanary = result.length > 1 && justNames.some(element => regexCanary.test(element)) ? true : false;
-//       if (result.length > 1) {
-//         if (anyNotRunning) {
-//           message = `This service has a canary version, and there may be a problem with either an instance of the canary version, or the production version. See "taskStatus" details below.
-
-// For more information on app performance, visit the prometheus and grafana dashboards you have configured with SENTINEL METRICS, or inspect app logs with SENTINEL INSPECT LOGS. You can also view system level metrics for your compute instances with SENTINEL CLUSTER INSPECT.`;
-//         } else {
-//           message = `This service has a canary version, and all instances of canary and production appear to be running. See "taskStatus" details below.
-
-// For more information on app performance, visit the prometheus and grafana dashboards you have configured with SENTINEL METRICS, or inspect app logs with SENTINEL INSPECT LOGS. You can also view system level metrics for your compute instances with SENTINEL CLUSTER INSPECT.`;
-//         }
-//       } else {
-//         if (anyNotRunning) {
-//           message = `This service has no canary version, but there may be a problem with the production version. See "taskStatus" details below.
-
-// For more information on app performance, visit the prometheus and grafana dashboards you have configured with SENTINEL METRICS, or inspect app logs with SENTINEL INSPECT LOGS.. You can also view system level metrics for your compute instances with SENTINEL CLUSTER INSPECT.`;
-//         } else {
-//           message = `This service has no canary version, and all instances appear to be running. See "taskStatus" details below.
-
-// For more information on app performance, visit the prometheus and grafana dashboards you have configured with SENTINEL METRICS, or inspect app logs with SENTINEL INSPECT LOGS. You can also view system level metrics for your compute instances with SENTINEL CLUSTER INSPECT.`;
-//         }
-//       }
-
-
-
-      let response = {
-        hasCanary,
-        message,
-        data: result
-      };
-      res.send(response);
-    } catch (err) {
-      console.error(err);
-    }
-  },
+module.exports = {
+  getServiceLogs,
+  listServices,
+  inspectService,
 
   async canaryDeploy(req, res, next) {
     const appName = req.params.appName;
