+++ models/httpError.js
@@ -1,0 +1,8 @@
+class HttpError extends Error {
+  constructor(error, errorCode) {
+    super(error);
+    this.code = errorCode;
+  }
+}
+
+module.exports = HttpError;
